メールのAPIに対処
