+++ backend/apps/mail/views/__init__.py
@@ -1,4 +1,3 @@
-from .mail_group_create_view import MailGroupCreateView
-from .mail_group_list_view import MailGroupListView
+from .mail_group_view import MailGroupView
 from .mail_group_detail_create_view import MailGroupDetailBulkCreateView
 from .mail_group_delete_view import MailGroupDeleteView
